# services/payments-svc/api/payments/v1/payment_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.7.1
// - protoc             v6.33.0
// source: api/payments/v1/payment.proto

package paymentsv1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationPaymentServiceApprovePayment = "/api.payments.v1.PaymentService/ApprovePayment"
const OperationPaymentServiceCancelPayment = "/api.payments.v1.PaymentService/CancelPayment"
const OperationPaymentServiceCreatePayment = "/api.payments.v1.PaymentService/CreatePayment"
const OperationPaymentServiceDeclinePayment = "/api.payments.v1.PaymentService/DeclinePayment"
const OperationPaymentServiceGetPayment = "/api.payments.v1.PaymentService/GetPayment"
const OperationPaymentServiceGetPaymentFileStatusReport = "/api.payments.v1.PaymentService/GetPaymentFileStatusReport"
const OperationPaymentServiceGetPaymentStatusReport = "/api.payments.v1.PaymentService/GetPaymentStatusReport"
const OperationPaymentServiceListParticipants = "/api.payments.v1.PaymentService/ListParticipants"
const OperationPaymentServiceListPayments = "/api.payments.v1.PaymentService/ListPayments"
const OperationPaymentServiceSubmitPaymentFile = "/api.payments.v1.PaymentService/SubmitPaymentFile"
const OperationPaymentServiceVerifyPayee = "/api.payments.v1.PaymentService/VerifyPayee"

type PaymentServiceHTTPServer interface {
	// ApprovePayment ApprovePayment records an approval. The maker can never approve its own
	// payment; each approver counts once.
	ApprovePayment(context.Context, *ApprovePaymentRequest) (*Payment, error)
	// CancelPayment CancelPayment withdraws a payment not yet approved (its maker), or pulls
	// a NORMAL payment from the netting queue before its cycle runs.
	CancelPayment(context.Context, *CancelPaymentRequest) (*Payment, error)
	// CreatePayment CreatePayment records a payment order from a maker.
	CreatePayment(context.Context, *CreatePaymentRequest) (*CreatePaymentResponse, error)
	// DeclinePayment DeclinePayment is an approver refusing a payment.
	DeclinePayment(context.Context, *DeclinePaymentRequest) (*Payment, error)
	// GetPayment GetPayment returns one payment with its audit trail.
	GetPayment(context.Context, *GetPaymentRequest) (*Payment, error)
	// GetPaymentFileStatusReport GetPaymentFileStatusReport renders a file's current pain.002.001.11.
	GetPaymentFileStatusReport(context.Context, *GetPaymentFileRequest) (*Document, error)
	// GetPaymentStatusReport GetPaymentStatusReport renders the payment's status as a pain.002.001.11.
	GetPaymentStatusReport(context.Context, *GetPaymentRequest) (*Document, error)
	// ListParticipants ListParticipants lists the banks on the network and whether each is live.
	ListParticipants(context.Context, *ListParticipantsRequest) (*ListParticipantsResponse, error)
	// ListPayments ListPayments lists the caller's organization's payments, newest first.
	ListPayments(context.Context, *ListPaymentsRequest) (*ListPaymentsResponse, error)
	// SubmitPaymentFile SubmitPaymentFile accepts a pain.001.001.10 customer credit transfer
	// initiation. Each transaction becomes a payment awaiting approval; the
	// answer carries the file-level pain.002.
	SubmitPaymentFile(context.Context, *SubmitPaymentFileRequest) (*PaymentFile, error)
	// VerifyPayee VerifyPayee asks the receiving bank whether an account is open and the
	// name matches its records.
	VerifyPayee(context.Context, *Party) (*PayeeCheck, error)
}

func RegisterPaymentServiceHTTPServer(s *http.Server, srv PaymentServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/payments", _PaymentService_CreatePayment0_HTTP_Handler(srv))
	r.GET("/v1/payments/{id}", _PaymentService_GetPayment0_HTTP_Handler(srv))
	r.GET("/v1/payments", _PaymentService_ListPayments0_HTTP_Handler(srv))
	r.POST("/v1/payments/{id}/approve", _PaymentService_ApprovePayment0_HTTP_Handler(srv))
	r.POST("/v1/payments/{id}/decline", _PaymentService_DeclinePayment0_HTTP_Handler(srv))
	r.POST("/v1/payments/{id}/cancel", _PaymentService_CancelPayment0_HTTP_Handler(srv))
	r.GET("/v1/payments/{id}/pain002", _PaymentService_GetPaymentStatusReport0_HTTP_Handler(srv))
	r.POST("/v1/payment-files", _PaymentService_SubmitPaymentFile0_HTTP_Handler(srv))
	r.GET("/v1/payment-files/{id}/pain002", _PaymentService_GetPaymentFileStatusReport0_HTTP_Handler(srv))
	r.POST("/v1/payees/verify", _PaymentService_VerifyPayee0_HTTP_Handler(srv))
	r.GET("/v1/network/participants", _PaymentService_ListParticipants0_HTTP_Handler(srv))
}

func _PaymentService_CreatePayment0_HTTP_Handler(srv PaymentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreatePaymentRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentServiceCreatePayment)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreatePayment(ctx, req.(*CreatePaymentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreatePaymentResponse)
		return ctx.Result(200, reply)
	}
}

func _PaymentService_GetPayment0_HTTP_Handler(srv PaymentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetPaymentRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentServiceGetPayment)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetPayment(ctx, req.(*GetPaymentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Payment)
		return ctx.Result(200, reply)
	}
}

func _PaymentService_ListPayments0_HTTP_Handler(srv PaymentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListPaymentsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentServiceListPayments)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListPayments(ctx, req.(*ListPaymentsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListPaymentsResponse)
		return ctx.Result(200, reply)
	}
}

func _PaymentService_ApprovePayment0_HTTP_Handler(srv PaymentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ApprovePaymentRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentServiceApprovePayment)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ApprovePayment(ctx, req.(*ApprovePaymentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Payment)
		return ctx.Result(200, reply)
	}
}

func _PaymentService_DeclinePayment0_HTTP_Handler(srv PaymentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeclinePaymentRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentServiceDeclinePayment)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeclinePayment(ctx, req.(*DeclinePaymentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Payment)
		return ctx.Result(200, reply)
	}
}

func _PaymentService_CancelPayment0_HTTP_Handler(srv PaymentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CancelPaymentRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentServiceCancelPayment)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CancelPayment(ctx, req.(*CancelPaymentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Payment)
		return ctx.Result(200, reply)
	}
}

func _PaymentService_GetPaymentStatusReport0_HTTP_Handler(srv PaymentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetPaymentRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentServiceGetPaymentStatusReport)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetPaymentStatusReport(ctx, req.(*GetPaymentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Document)
		return ctx.Result(200, reply)
	}
}

func _PaymentService_SubmitPaymentFile0_HTTP_Handler(srv PaymentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SubmitPaymentFileRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentServiceSubmitPaymentFile)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SubmitPaymentFile(ctx, req.(*SubmitPaymentFileRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PaymentFile)
		return ctx.Result(200, reply)
	}
}

func _PaymentService_GetPaymentFileStatusReport0_HTTP_Handler(srv PaymentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetPaymentFileRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentServiceGetPaymentFileStatusReport)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetPaymentFileStatusReport(ctx, req.(*GetPaymentFileRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Document)
		return ctx.Result(200, reply)
	}
}

func _PaymentService_VerifyPayee0_HTTP_Handler(srv PaymentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in Party
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentServiceVerifyPayee)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.VerifyPayee(ctx, req.(*Party))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PayeeCheck)
		return ctx.Result(200, reply)
	}
}

func _PaymentService_ListParticipants0_HTTP_Handler(srv PaymentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListParticipantsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentServiceListParticipants)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListParticipants(ctx, req.(*ListParticipantsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListParticipantsResponse)
		return ctx.Result(200, reply)
	}
}

type PaymentServiceHTTPClient interface {
	ApprovePayment(ctx context.Context, req *ApprovePaymentRequest, opts ...http.CallOption) (rsp *Payment, err error)
	CancelPayment(ctx context.Context, req *CancelPaymentRequest, opts ...http.CallOption) (rsp *Payment, err error)
	CreatePayment(ctx context.Context, req *CreatePaymentRequest, opts ...http.CallOption) (rsp *CreatePaymentResponse, err error)
	DeclinePayment(ctx context.Context, req *DeclinePaymentRequest, opts ...http.CallOption) (rsp *Payment, err error)
	GetPayment(ctx context.Context, req *GetPaymentRequest, opts ...http.CallOption) (rsp *Payment, err error)
	GetPaymentFileStatusReport(ctx context.Context, req *GetPaymentFileRequest, opts ...http.CallOption) (rsp *Document, err error)
	GetPaymentStatusReport(ctx context.Context, req *GetPaymentRequest, opts ...http.CallOption) (rsp *Document, err error)
	ListParticipants(ctx context.Context, req *ListParticipantsRequest, opts ...http.CallOption) (rsp *ListParticipantsResponse, err error)
	ListPayments(ctx context.Context, req *ListPaymentsRequest, opts ...http.CallOption) (rsp *ListPaymentsResponse, err error)
	SubmitPaymentFile(ctx context.Context, req *SubmitPaymentFileRequest, opts ...http.CallOption) (rsp *PaymentFile, err error)
	VerifyPayee(ctx context.Context, req *Party, opts ...http.CallOption) (rsp *PayeeCheck, err error)
}

type PaymentServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewPaymentServiceHTTPClient(client *http.Client) PaymentServiceHTTPClient {
	return &PaymentServiceHTTPClientImpl{client}
}

func (c *PaymentServiceHTTPClientImpl) ApprovePayment(ctx context.Context, in *ApprovePaymentRequest, opts ...http.CallOption) (*Payment, error) {
	var out Payment
	pattern := "/v1/payments/{id}/approve"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPaymentServiceApprovePayment))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *PaymentServiceHTTPClientImpl) CancelPayment(ctx context.Context, in *CancelPaymentRequest, opts ...http.CallOption) (*Payment, error) {
	var out Payment
	pattern := "/v1/payments/{id}/cancel"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPaymentServiceCancelPayment))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *PaymentServiceHTTPClientImpl) CreatePayment(ctx context.Context, in *CreatePaymentRequest, opts ...http.CallOption) (*CreatePaymentResponse, error) {
	var out CreatePaymentResponse
	pattern := "/v1/payments"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPaymentServiceCreatePayment))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *PaymentServiceHTTPClientImpl) DeclinePayment(ctx context.Context, in *DeclinePaymentRequest, opts ...http.CallOption) (*Payment, error) {
	var out Payment
	pattern := "/v1/payments/{id}/decline"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPaymentServiceDeclinePayment))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *PaymentServiceHTTPClientImpl) GetPayment(ctx context.Context, in *GetPaymentRequest, opts ...http.CallOption) (*Payment, error) {
	var out Payment
	pattern := "/v1/payments/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPaymentServiceGetPayment))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *PaymentServiceHTTPClientImpl) GetPaymentFileStatusReport(ctx context.Context, in *GetPaymentFileRequest, opts ...http.CallOption) (*Document, error) {
	var out Document
	pattern := "/v1/payment-files/{id}/pain002"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPaymentServiceGetPaymentFileStatusReport))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *PaymentServiceHTTPClientImpl) GetPaymentStatusReport(ctx context.Context, in *GetPaymentRequest, opts ...http.CallOption) (*Document, error) {
	var out Document
	pattern := "/v1/payments/{id}/pain002"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPaymentServiceGetPaymentStatusReport))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *PaymentServiceHTTPClientImpl) ListParticipants(ctx context.Context, in *ListParticipantsRequest, opts ...http.CallOption) (*ListParticipantsResponse, error) {
	var out ListParticipantsResponse
	pattern := "/v1/network/participants"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPaymentServiceListParticipants))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *PaymentServiceHTTPClientImpl) ListPayments(ctx context.Context, in *ListPaymentsRequest, opts ...http.CallOption) (*ListPaymentsResponse, error) {
	var out ListPaymentsResponse
	pattern := "/v1/payments"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPaymentServiceListPayments))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *PaymentServiceHTTPClientImpl) SubmitPaymentFile(ctx context.Context, in *SubmitPaymentFileRequest, opts ...http.CallOption) (*PaymentFile, error) {
	var out PaymentFile
	pattern := "/v1/payment-files"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPaymentServiceSubmitPaymentFile))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *PaymentServiceHTTPClientImpl) VerifyPayee(ctx context.Context, in *Party, opts ...http.CallOption) (*PayeeCheck, error) {
	var out PayeeCheck
	pattern := "/v1/payees/verify"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPaymentServiceVerifyPayee))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}