# services/payments-svc/internal/server/openapi.yaml
...
openapi: 3.0.3
info:
    title: Reserve interbank settlement payment hub API
    version: 1.0.0
paths:
...
    /v1/ops/payments:
        get:
            tags:
                - BankOperationsService
            description: ListBankPayments lists the bank's clients' payments, with staff detail.
            operationId: BankOperationsService_ListBankPayments
            parameters:
                - name: status
                  in: query
                  schema:
                    type: string
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageToken
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.ListBankPaymentsResponse'
    /v1/ops/payments/{id}:
        get:
            tags:
                - BankOperationsService
            description: GetBankPayment returns one of the bank's clients' payments, with staff detail.
            operationId: BankOperationsService_GetBankPayment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
    /v1/payees/verify:
        post:
            tags:
                - PaymentService
            description: |-
                VerifyPayee asks the receiving bank whether an account is open and the
                 name matches its records.
            operationId: PaymentService_VerifyPayee
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.Party'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.PayeeCheck'
    /v1/payment-files:
        post:
            tags:
                - PaymentService
            description: |-
                SubmitPaymentFile accepts a pain.001.001.10 customer credit transfer
                 initiation. Each transaction becomes a payment awaiting approval; the
                 answer carries the file-level pain.002.
            operationId: PaymentService_SubmitPaymentFile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.SubmitPaymentFileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.PaymentFile'
    /v1/payment-files/{id}/pain002:
        get:
            tags:
                - PaymentService
            description: GetPaymentFileStatusReport renders a file's current pain.002.001.11.
            operationId: PaymentService_GetPaymentFileStatusReport
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Document'
    /v1/payments:
        get:
            tags:
                - PaymentService
            description: ListPayments lists the caller's organization's payments, newest first.
            operationId: PaymentService_ListPayments
            parameters:
                - name: status
                  in: query
                  schema:
                    type: string
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageToken
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.ListPaymentsResponse'
        post:
            tags:
                - PaymentService
            description: CreatePayment records a payment order from a maker.
            operationId: PaymentService_CreatePayment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.CreatePaymentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.CreatePaymentResponse'
    /v1/payments/{id}:
        get:
            tags:
                - PaymentService
            description: GetPayment returns one payment with its audit trail.
            operationId: PaymentService_GetPayment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
    /v1/payments/{id}/approve:
        post:
            tags:
                - PaymentService
            description: |-
                ApprovePayment records an approval. The maker can never approve its own
                 payment; each approver counts once.
            operationId: PaymentService_ApprovePayment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.ApprovePaymentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
    /v1/payments/{id}/cancel:
        post:
            tags:
                - PaymentService
            description: |-
                CancelPayment withdraws a payment not yet approved (its maker), or pulls
                 a NORMAL payment from the netting queue before its cycle runs.
            operationId: PaymentService_CancelPayment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.CancelPaymentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
    /v1/payments/{id}/decline:
        post:
            tags:
                - PaymentService
            description: DeclinePayment is an approver refusing a payment.
            operationId: PaymentService_DeclinePayment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.DeclinePaymentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
    /v1/payments/{id}/pain002:
        get:
            tags:
                - PaymentService
            description: GetPaymentStatusReport renders the payment's status as a pain.002.001.11.
            operationId: PaymentService_GetPaymentStatusReport
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Document'
...
components:
    schemas:
...
        api.payments.v1.ApprovePaymentRequest:
            type: object
            properties:
                id:
                    type: string
...
        api.payments.v1.CreatePaymentRequest:
            type: object
            properties:
                debtorAccount:
                    type: string
                creditor:
                    $ref: '#/components/schemas/api.payments.v1.Party'
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                priority:
                    type: string
                remittance:
                    type: string
                endToEndId:
                    type: string
        api.payments.v1.CreatePaymentResponse:
            type: object
            properties:
                payment:
                    $ref: '#/components/schemas/api.payments.v1.Payment'
                duplicateRequest:
                    type: boolean
                    description: |-
                        True when the Idempotency-Key matched an earlier request with the same
                         body: nothing new was created.
...
        api.payments.v1.Document:
            type: object
            properties:
                messageType:
                    type: string
                xml:
                    type: string
            description: Document is an ISO 20022 message rendered as XML.
...
        api.payments.v1.ListBankPaymentsResponse:
            type: object
            properties:
                payments:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.payments.v1.Payment'
                nextPageToken:
                    type: string
...
        api.payments.v1.ListPaymentsResponse:
            type: object
            properties:
                payments:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.payments.v1.Payment'
                nextPageToken:
                    type: string
...
        api.payments.v1.Party:
            type: object
            properties:
                name:
                    type: string
                routingNumber:
                    type: string
                account:
                    type: string
            description: Party identifies a payee by name, ABA routing number and account.
        api.payments.v1.PayeeCheck:
            type: object
            properties:
                accountStatus:
                    type: string
                nameMatch:
                    type: string
                registeredName:
                    type: string
            description: |-
                PayeeCheck is the receiving bank's answer before any money moves, as the
                 Fed's Payee Name Verification and banks' account-validation services give
                 it.
        api.payments.v1.Payment:
            type: object
            properties:
                id:
                    type: string
                endToEndId:
                    type: string
                uetr:
                    type: string
                idempotencyKey:
                    type: string
                organizationId:
                    type: string
                debtorAccount:
                    type: string
                creditor:
                    $ref: '#/components/schemas/api.payments.v1.Party'
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                priority:
                    type: string
                remittance:
                    type: string
                status:
                    type: string
                    description: |-
                        AWAITING_APPROVAL | IN_PROCESS | ON_HOLD | AWAITING_LIQUIDITY |
                         QUEUED_FOR_NETTING | SETTLED | REJECTED | CANCELLED | EXPIRED | BLOCKED
                isoStatus:
                    type: string
                    description: RCVD | ACTC | PATC | ACCP | PDNG | ACSP | ACSC | ACCC | RJCT | CANC | BLCK
                reasonCode:
                    type: string
                route:
                    type: string
                    description: instant, 24x7 | netting cycle | book transfer, same bank
                routeNote:
                    type: string
                payeeCheck:
                    $ref: '#/components/schemas/api.payments.v1.PayeeCheck'
                approvalsRequired:
                    type: integer
                    format: int32
                approvals:
                    type: array
                    items:
                        type: string
                createdBy:
                    type: string
                createdAt:
                    type: string
                    format: date-time
                settledAt:
                    type: string
                    format: date-time
                ofacReportDue:
                    type: string
                    format: date-time
                history:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.payments.v1.PaymentEvent'
                fileId:
                    type: string
                sanctionsMatch:
                    type: string
            description: Payment is a payment order and its lifecycle.
...
        api.payments.v1.PaymentFile:
            type: object
            properties:
                id:
                    type: string
                messageId:
                    type: string
                numberOfTransactions:
                    type: integer
                    format: int32
                controlSum:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                payments:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.payments.v1.Payment'
                rejections:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.payments.v1.FileRejection'
                    description: Transactions refused at validation, by end-to-end id, with a reason.
                statusReport:
                    $ref: '#/components/schemas/api.payments.v1.Document'
                duplicateRequest:
                    type: boolean
...
        api.payments.v1.SubmitPaymentFileRequest:
            type: object
            properties:
                xml:
                    type: string
...